+++ main.py
@@ -7,7 +7,7 @@ app = Flask(__name__)
 @app.route("/")
 def checker():
     checker = Checker(request.args)
-#    return checker.make_report_0()
+    #    return checker.make_report_0()
     results = checker.do_checks()
     report = checker.make_report(results)
     return report
