+++ templates/header.html
@@ -11,4 +11,5 @@
   {% else %}
   <p>Non-verbose report. Add &verbose=1 URL parameter to include info about tests passed.</p>
   {% endif %}
+<p>^ verbose/non-verbose option not working</p>
 </header>
